+++ packages/koot/React/inject/_cache-keys.js
@@ -1,0 +1,8 @@
+/**
+ * inject 缓存键值
+ */
+module.exports = {
+    scriptsRunFirst: 'SCRIPTS_RUN_FIRST',
+    scriptsInBody: 'SCRIPTS_IN_BODY',
+    uriServiceWorker: 'URI_SERVICE_WORKER'
+};