+++ packages/koot/React/inject/scripts.js
@@ -1,10 +1,25 @@
 const path = require('path');
 
-const { chunkNameClientRunFirst } = require('../../defaults/before-build');
+const {
+    chunkNameClientRunFirst,
+    scriptTagEntryAttributeName
+} = require('../../defaults/before-build');
 const defaultEntrypoints = require('../../defaults/entrypoints');
 const readClientFile = require('../../utils/read-client-file');
 const getClientFilePath = require('../../utils/get-client-file-path');
 const getSSRStateString = require('../../libs/get-ssr-state-string');
+const {
+    scriptsRunFirst,
+    scriptsInBody,
+    uriServiceWorker
+} = require('./_cache-keys');
+
+/**
+ * @type {number} run-first 入口的文件尺寸阈值
+ * - 如果超过这个值，会采用 <script> 引用的方式
+ * - 如果小于等于这个值，直接将文件内容写入 HTML
+ */
+const thresholdRunFirst = 10 * 1000;
 
 /**
  * 注入: JavaScript 代码
@@ -33,7 +48,35 @@ module.exports = ({
     );
     const isProd = !isDev;
 
-    if (isDev || typeof injectCache.scriptsInBody === 'undefined') {
+    if (isDev || typeof injectCache[scriptsRunFirst] === 'undefined') {
+        const filename = `${chunkNameClientRunFirst}.js`;
+        const name = '*run-first';
+        if (isDev) {
+            injectCache[scriptsRunFirst] = combineFilePaths(
+                name,
+                filename,
+                localeId
+            );
+        } else {
+            const content = scriptTagsFromContent(
+                name,
+                filename,
+                localeId,
+                compilation
+            );
+            if (content.length > thresholdRunFirst) {
+                injectCache[scriptsRunFirst] = combineFilePaths(
+                    name,
+                    filename,
+                    localeId
+                );
+            } else {
+                injectCache[scriptsRunFirst] = content;
+            }
+        }
+    }
+
+    if (isDev || typeof injectCache[scriptsInBody] === 'undefined') {
         let r = '';
 
         // 入口: critical
@@ -43,12 +86,9 @@ module.exports = ({
                 .map(file => {
                     if (isDev) {
                         // return `<script type="text/javascript" src="${getClientFilePath(true, file)}"></script>`;
-                        return combineFilePaths(true, file);
+                        return combineFilePaths('critical', true, file);
                     }
-                    return `<script type="text/javascript">${readClientFile(
-                        true,
-                        file
-                    )}</script>`;
+                    return scriptTagsFromContent('critical', true, file);
                 })
                 .join('');
         }
@@ -68,7 +108,7 @@ module.exports = ({
                         //     true,
                         //     file
                         // )}" defer></script>`;
-                        return combineFilePaths(true, file);
+                        return combineFilePaths(key, true, file);
                     })
                     .join('');
             }
@@ -82,15 +122,17 @@ module.exports = ({
         if (
             pwaAuto &&
             (process.env.WEBPACK_BUILD_TYPE === 'spa' ||
-                typeof injectCache.pathnameSW === 'string')
+                typeof injectCache[uriServiceWorker] === 'string')
         ) {
             r += `<script id="__koot-pwa-register-sw" type="text/javascript">`;
             if (isProd) {
                 r +=
                     `if ('serviceWorker' in navigator) {` +
                     `window.addEventListener('load', function() {` +
-                    // + `navigator.serviceWorker.register("${injectCache.pathnameSW}?koot=${process.env.KOOT_VERSION}",`
-                    `navigator.serviceWorker.register("${injectCache.pathnameSW ||
+                    // + `navigator.serviceWorker.register("${injectCache[uriServiceWorker]}?koot=${process.env.KOOT_VERSION}",`
+                    `navigator.serviceWorker.register("${injectCache[
+                        uriServiceWorker
+                    ] ||
                         JSON.parse(
                             process.env.KOOT_PWA_PATHNAME
                         )}?koot=0.11",` +
@@ -105,7 +147,7 @@ module.exports = ({
             r += `</script>`;
         }
 
-        injectCache.scriptsInBody = r;
+        injectCache[scriptsInBody] = r;
     }
 
     return (
@@ -114,48 +156,38 @@ module.exports = ({
         `window.__KOOT_LOCALEID__ = "${SSRState.localeId || ''}";` +
         `window.__KOOT_SSR_STATE__ = ${getSSRStateString(SSRState)};` +
         `</script>` +
-        getClientRunFirstJS(localeId, compilation) +
-        `${injectCache.scriptsInBody}`
+        // getClientRunFirstJS(localeId, compilation) +
+        injectCache[scriptsRunFirst] +
+        injectCache[scriptsInBody]
     );
 };
 
 /**
- * 客户端预先执行 JS 的代码
- * @param {*} localeId
- * @param {*} compilation
- * @returns {String}
- */
-const getClientRunFirstJS = (localeId, compilation) => {
-    const filename = `${chunkNameClientRunFirst}.js`;
-
-    if (process.env.WEBPACK_BUILD_ENV === 'dev') {
-        return combineFilePaths(filename, localeId);
-        // return `<script type="text/javascript" src="${getClientFilePath(
-        //     filename,
-        //     localeId
-        // )}"></script>`;
-    }
-
-    return `<script type="text/javascript">${readClientFile(
-        filename,
-        localeId,
-        compilation
-    )}</script>`;
-};
-
-/**
- * 返回 script 标签
+ * 返回引用地址的 script 标签
  * 如果有多个结果，会返回包含多个标签的 HTML 结果
+ * @param {string} name 入口名/标识符
  * @param {...any} args `utils/get-client-file-path` 对应的参数
  * @returns {String} 整合的 HTML 结果
  */
-const combineFilePaths = (...args) => {
+const combineFilePaths = (name, ...args) => {
     let pathnames = getClientFilePath(...args);
     if (!Array.isArray(pathnames)) pathnames = [pathnames];
     return pathnames
         .map(
             pathname =>
-                `<script type="text/javascript" src="${pathname}" defer></script>`
+                `<script type="text/javascript" src="${pathname}" defer ${scriptTagEntryAttributeName}="${name}"></script>`
         )
         .join('');
 };
+
+/**
+ * 返回 script 标签，标签内容为源代码字符串
+ * 如果有多个结果，会返回包含多个标签的 HTML 结果
+ * @param {string} name 入口名/标识符
+ * @param {...any} args `utils/read-client-file` 对应的参数
+ * @returns {String} 整合的 HTML 结果
+ */
+const scriptTagsFromContent = (name, ...args) =>
+    `<script type="text/javascript" ${scriptTagEntryAttributeName}="${name}">${readClientFile(
+        ...args
+    )}</script>`;