+++ CHANGELOG.md
@@ -6,12 +6,17 @@
 
 -   **重大改动**
     -   _Node.js_ 最低版本要求提升到 `8.12.0`
-    -   [SSR & 客户端渲染] 组件 CSS 的 `<style>` 标签上不再有 `id` 属性，以避免和元素冲突
-        -   如果项目中有用到根据 `id` 选择 `<style>` 标签的场景，可改为选择标签属性 `[data-koot-module]`
+    -   [SSR & 客户端渲染 & SPA 模板生成]
+        -   组件 CSS 的 `<style>` 标签上不再有 `id` 属性，以避免和元素冲突
+            -   如果项目中有用到根据 `id` 选择 `<style>` 标签的场景，可改为选择标签属性 `[data-koot-module]`
+        -   特殊 JS 入口 _CLIENT_RUN_FIRST_ 如果文件尺寸大于 10KB，会改为引用请求的方式加载
+        -   特殊 JS 入口 _CLIENT_RUN_FIRST_ 现默认引入 `regenerator-runtime/runtime`
+            -   现在 _Koot.js_ 项目默认支持在浏览器环境中使用 _async_ / _await_ 开发
 -   **新特性**
     -   **新工具函数** `koot/utils/client-get-styles` - 获取当前全局 CSS 和所有组件 CSS。具体用法请参见文档 [全局与工具函数/工具函数/客户端](https://koot.js.org/#/utilities?id=客户端) ([#185](https://github.com/cmux/koot/issues/185))
 -   优化
-    -   [服务器] 优化服务器代码的文件尺寸 ([#172](https://github.com/cmux/koot/issues/172))
+    -   [服务器] 优化服务器代码的文件尺寸 ([#172](https://github.com/cmux/koot/issues/172), [#186](https://github.com/cmux/koot/issues/186))
+    -   [SSR & 客户端渲染 & SPA 模板生成] 所有注入/插入的 `<script>` 标签现在均会新增 `[data-koot-entry]` 属性
 
 ## [0.11.15] - 2019-10-15